Started refactoring RTF stuff into functions instead of classes
