+++ extract_msg/_rtf/__init__.py
@@ -1,0 +1,7 @@
+"""
+Module that provides access to functions to help manage RTF data.
+"""
+
+
+from .token import Token, TokenType
+from .tokenize_rtf import tokenizeRTF
